pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Python Fast Api for the Decision Optimization Project"
authors = ["EinarSalbouvet <esalo@equinor.com>"]
license = "GNU GENERAL PUBLIC LICENSE"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
alembic = "^1.16.5"
sqlalchemy = "^2.0.43"
fastapi = "^0.116.1"
aiosqlite = "^0.21.0"
uvicorn = "^0.35.0"
pytest = "^8.4.2"
httpx = "^0.28.1"
pytest-asyncio = "^1.0.0"
requests = "^2.32.5"
authlib = "^1.6.3"
jwt = "^1.3.1"
pytest-cov = "^6.1.1"
pydantic-settings = "^2.9.1"
aioodbc = "^0.5.0"
aiomysql = "^0.2.0"
azure-identity = "^1.24.0"
aiohttp = "^3.12.15"
urllib3 ="2.5.0"
odata-query = "^0.10.0"
greenlet = "^3.2.4"
python-dotenv = "^1.1.1"
asgi-lifespan = "^2.1.0"
pyagrum = "^2.2.1"
black = "^25.9.0"

[tool.pytest.ini_options]
asyncio_mode = "strict"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
